refactor: factorize common code across files checkUserIsValidOrThrowError
