+++ src/server/queueing/workers/send-validation-email.ts
@@ -2,23 +2,19 @@ import { isAfter } from "date-fns/isAfter";
 import { isBefore } from "date-fns/isBefore";
 import PgBoss from "pg-boss";
 
-import { db } from "@/lib/kysely";
+import { checkUserIsValidOrThrowError } from "./utils";
 import { getIncubator } from "@/lib/kysely/queries/incubators";
 import { getUsersByIncubatorId } from "@/lib/kysely/queries/teams";
-import { getUserBasicInfo, getUserStartups } from "@/lib/kysely/queries/users";
+import { getUserStartups } from "@/lib/kysely/queries/users";
 import {
     SendNewMemberValidationEmailSchema,
     SendNewMemberValidationEmailSchemaType,
 } from "@/models/jobs/member";
-import {
-    incubatorToModel,
-    memberPublicInfoToModel,
-    userStartupToModel,
-} from "@/models/mapper";
+import { incubatorToModel, userStartupToModel } from "@/models/mapper";
 import config from "@/server/config";
 import { sendEmail } from "@/server/config/email.config";
 import { EMAIL_TYPES } from "@/server/modules/email";
-import { BusinessError, NoDataError } from "@/utils/error";
+import { BusinessError } from "@/utils/error";
 
 export const sendNewMemberValidationEmailTopic =
     "send-new-member-validation-email";
@@ -27,25 +23,8 @@ export async function sendNewMemberValidationEmail(
     job: PgBoss.Job<SendNewMemberValidationEmailSchemaType>
 ) {
     const data = SendNewMemberValidationEmailSchema.parse(job.data);
+    await checkUserIsValidOrThrowError(data.userId);
     const now = new Date();
-    const memberDbData = await getUserBasicInfo({ uuid: data.userId });
-    if (!memberDbData) {
-        throw new NoDataError(
-            `Pas de membre trouvé pour l'id : ${data.userId}`
-        );
-    }
-    const newMember = memberPublicInfoToModel(memberDbData);
-    const userMissions = await db
-        .selectFrom("missions")
-        .selectAll()
-        .where("user_id", "=", data.userId)
-        .execute();
-    if (!userMissions.length) {
-        throw new BusinessError(
-            "NoActiveMissionForUser",
-            `User ${data.userId} does not have any missions`
-        );
-    }
     const userStartups = (await getUserStartups(data.userId)).filter(
         (startup) => {
             return (
